Move vFFA generation in a specific file
