+++ zgoubidoo/fieldmaps/__init__.py
@@ -1,2 +1,3 @@
 """Field map manipulations module."""
-from .fieldmap import EngeModel, CartesianFieldMap, PolarFieldMap, VFFAFieldMap
+from .fieldmap import EngeModel, CartesianFieldMap, PolarFieldMap
+from .vffa import VFFAFieldMap
